requirements.txt
# Photo Organizer - Dependências do Projeto
# Python >= 3.9

# === Processamento de Imagens ===
Pillow>=10.0.0              # Manipulação básica de imagens e EXIF
piexif>=1.1.3               # Leitura/escrita de metadados EXIF completos
imagehash>=4.3.1            # Hashes perceptuais para detectar duplicatas similares

# === Configuração ===
PyYAML>=6.0                 # Leitura de arquivos de configuração YAML

# === Interface e Feedback ===
tqdm>=4.65.0                # Barras de progresso no terminal

# === Detecção de Tipo de Arquivo (Windows) ===
python-magic-bin>=0.4.14    # Validação de tipos de arquivo (versão Windows)

# === Utilitários ===
colorama>=0.4.6             # Cores no terminal Windows

# === Persistência e DB ===
sqlite-utils>=3.20.0        # Utilitários opcionais para SQLite (leve)

# === Desenvolvimento (opcional, mas recomendado) ===
pytest>=7.4.0               # Framework de testes
black>=23.0.0               # Formatador de código
pylint>=2.17.0              # Linter para qualidade de código

# Pin heavier numeric libs only if needed later
# numpy and scipy are optional; add when using imagehash or advanced analysis